% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifie_type_variable.R
\name{identifie_type_variable}
\alias{identifie_type_variable}
\title{Une fonction auxiliaires qui reconnait certains types de variables dans une table, et les renomme en conséquence}
\usage{
identifie_type_variable(tab, vars = names(tab))
}
\arguments{
\item{tab}{une table de données}

\item{vars}{un vecteur de noms de variables à analyser (par défaut, toutes les variables de la table 'tab')}
}
\value{
une table analogue à celle entrée en input, mais dont certaines variables auront éventuellement été renommées
}
\description{
A partir d'une table de données en entrée, cette fonction repère si les variables de la table
correspondent à certains types de donnée en particulier (année, sexe, territoire, etc) et,
si c'est bien le cas, change le nom de ces variables pour correspondre à ce type.
}
