% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liste_datadrees.R
\name{liste_datadrees}
\alias{liste_datadrees}
\title{Récupère la liste de tous les fichiers data.drees}
\usage{
liste_datadrees(rubriques = c(), producteurs = c())
}
\arguments{
\item{rubriques}{une liste de rubriques (par défaut toutes les rubriques disponibles)}

\item{producteurs}{une liste de producteurs (par défaut tous les producteurs disponibles)}
}
\value{

}
\description{
La fonction retourne un tableau contenant le nom et les métadonnées de tous les fichiers data.drees
diffusés par la DREES. Ces fichiers sont plus nombreux que les jeux de données, car un même
jeu de données peut contenir plusieurs fichiers (.xlsx, .csv, .zip, etc.)
}
\details{
Si une liste de rubriques ou de producteurs est indiquée, seuls les fichiers de ces rubriques ou producteurs seront retenus.
}
\examples{
liste_datadrees(rubriques = c("retraite"))
liste_datadrees(producteurs = c("OSOL"))
}
