% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettoie_table.R
\name{nettoie_table}
\alias{nettoie_table}
\title{Met en forme une table de donnée}
\usage{
nettoie_table(tab, forme_longue = TRUE)
}
\arguments{
\item{tab}{une table de données extraites d'un onglet d'un fichier excel data.drees}

\item{forme_longue}{indique si les intitulés de variables en colonnes doivent être remis en lignes}
}
\value{
une table de données plus propre
}
\description{
Cette fonction traite et nettoie une table de données lue dans un onglet d'un fichier Excel
diffusé sous data.drees. Elle repère les informations correspondant aux intitulés des colonnes
et celle correspondant à des variables de catégories.
}
\details{
Cette fonction auxiliaire est notamment utilisée dans la fonction extrait_onglet.
}
