% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OSOLdatadrees}
\alias{OSOLdatadrees}
\title{Base recensant les fichiers data.Drees de la sous-direction de l'observation de la solidarité de la DREES}
\format{
Un data frame avec 266 observations et 40 variables
}
\source{
\url{https://data.drees.solidarites-sante.gouv.fr/pages/accueil/}
}
\usage{
OSOLdatadrees
}
\description{
La base contient toutes les métadonnées pour la liste complète des fichiers data.drees (fichiers .xlsx, .xls, .zip, .7z, etc)
contenus dans les jeux de données produits par la sous-direction de l'observation de la solidarité (OSOL) de la DREES
}
\details{
La base actuelle a été créée le 7 février 2021. Elle n'est pas à jour des jeux de données créés ou mis à jour après cette date
}
\keyword{datasets}
