% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathdatadrees.R
\name{pathdatadrees}
\alias{pathdatadrees}
\title{Fonction auxiliaire retournant le chemin d'accès d'un fichier à télécharger sous data.drees}
\usage{
pathdatadrees(jdd, fichier)
}
\arguments{
\item{jdd}{le nom du jeu de données}

\item{fichier}{le nom du fichier (.xlsx, .csv, etc.)}
}
\value{
le chemin d'accès du fichier sous data.Drees
}
\description{
Fonction auxiliaire retournant le chemin d'accès d'un fichier à télécharger sous data.drees
}
\examples{
pathdatadrees("549_la-prestation-de-compensation-du-handicap-pch","pch_et_actp_beneficiaires_par_sexe_et_age_montants_verses_donnees_2016.xlsx")
}
